Fix threading dispatch error, build db controller
